+++ R/model-options.R
@@ -43,7 +43,20 @@ read_options <- function(type) {
 #' @return TRUE if valid otherwise throw an error
 #' @export
 validate_model_options <- function(data, options) {
-  ## This must return TRUE if valid - otherwise throw an error
-  ## TODO: mrc-795 Add real validation
+
+  required_options <- c("area_scope", "area_level",
+                        "calendar_quarter_t1", "calendar_quarter_t2",
+                        "survey_prevalence")
+
+  if(!all(required_options %in% names(options)))
+    stop(paste("Required model options not supplied:",
+               paste(setdiff(required_options, names(options)), collapse = ", ")))
+
+  ## TODO: better approach to check file exists and is valid?
+  if(is.null(data$art_number) &&
+     (!is.null(options$include_art_t1) && options$include_art_t1 == "true" ||
+      !is.null(options$include_art_t2) && options$include_art_t2 == "true"))
+    stop("ART dataset not provided. ART data cannot be selected Yes to include.")
+  
   TRUE
 }